Update 07-01-24 @10:51
Finished video ProfileView - Refactor - ViewModel
