+++ MyDubDubGrub/Views/Screens/ProfileView/ProfileView.swift
@@ -1,0 +1,125 @@
+//
+//  ProfileView.swift
+//  MyDubDubGrub
+//
+//  Created by Mark Perryman on 6/29/24.
+//
+
+import SwiftUI
+import CloudKit
+
+struct ProfileView: View {
+	// MARK: - State Object
+	@StateObject private var viewModel = ProfileViewModel()
+	
+    var body: some View {
+		VStack {
+			ZStack {
+				NameBackgroundView()
+				
+				HStack(spacing: 16) {
+					ZStack {
+						AvatarView(image: viewModel.avatar, size: 84)
+						EditImage()
+					}
+					.padding(.leading, 12)
+					.onTapGesture {
+						viewModel.isShowingPhotoPicker = true
+					}
+
+					VStack(spacing: 1) {
+						TextField("First Name", text: $viewModel.firstName)
+							.profileNameStyle()
+						TextField("Last Name", text: $viewModel.lastName)
+							.profileNameStyle()
+						TextField("Company Name", text: $viewModel.companyName)
+					}
+					.padding(.trailing, 16)
+				}
+				.padding()
+			}
+			
+			VStack(alignment: .leading, spacing: 8) {
+				CharactersRemainView(currentCount: viewModel.bio.count)
+				TextEditor(text: $viewModel.bio)
+					.frame(height: 100)
+					.overlay(
+						RoundedRectangle(cornerRadius: 8)
+							.stroke(Color.secondary, lineWidth: 1.0)
+					)
+			}
+			.padding(.horizontal)
+			
+			Spacer()
+			
+			Button {
+//				createProfile()
+			} label: {
+				DDGButton(title: "Create Profile")
+			}
+			.padding(.bottom)
+		}
+		.navigationTitle("Profile")
+		.toolbar {
+			Button {
+				dismissKeyboard()
+			} label: {
+				Image(systemName: "keyboard.chevron.compact.down")
+			}
+		}
+		.onAppear {
+			viewModel.getProfile()
+		}
+		.alert(item: $viewModel.alertItem, content: { alertItem in
+			Alert(title: alertItem.title, message: alertItem.message, dismissButton: alertItem.dismissButton)
+		})
+		.sheet(isPresented: $viewModel.isShowingPhotoPicker) {
+			PhotoPicker(image: $viewModel.avatar)
+		}
+    }	
+}
+
+#Preview {
+	NavigationView {
+		ProfileView()
+	}
+}
+
+// MARK: - Sub Views
+struct NameBackgroundView: View {
+	var body: some View {
+		Color(.secondarySystemBackground)
+			.frame(height: 130)
+			.cornerRadius(12)
+			.padding(.horizontal)
+	}
+}
+
+struct EditImage: View {
+	var body: some View {
+		Image(systemName: "square.and.pencil")
+			.resizable()
+			.scaledToFit()
+			.frame(width: 14, height: 14)
+			.foregroundColor(.white)
+			.offset(y: 30)
+	}
+}
+
+struct CharactersRemainView: View {
+	// MARK: - Properties
+	var currentCount: Int
+	
+	var body: some View {
+		Text("Bio: ")
+			.font(.callout)
+			.foregroundColor(.secondary)
+		+ Text("\(100 - currentCount)")
+			.font(.callout)
+			.bold()
+			.foregroundColor(currentCount <= 100 ? .brandPrimary : Color(.systemPink))
+		+ Text(" Characters Remain")
+			.font(.callout)
+			.foregroundColor(.secondary)
+	}
+}
